refactor: pages
